% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{append_interactions}
\alias{append_interactions}
\title{Add Interactions to Predictor Matrix}
\usage{
append_interactions(x, interactions = NULL)
}
\arguments{
\item{x}{The original matrix containing the source of the interaction terms.}

\item{interactions}{An n_interactions x 2 matrix containing interactions
to add to the original matrix x.}
}
\value{
The matrix x with new columns associated wtih each interaction.
}
\description{
Add Interactions to Predictor Matrix
}
