% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_generation.R
\name{counterfactual_ts}
\alias{counterfactual_ts}
\title{Generate Counterfactual versions of a ts_inter object}
\usage{
counterfactual_ts(ts, w0, w1, start_ix = NULL)
}
\description{
Generate Counterfactual versions of a ts_inter object
}
