% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPV.R
\name{extractPV}
\alias{extractPV}
\title{Function to extract p-values from a permutation test dataset}
\usage{
extractPV(permSTAT, modelratio, trm = 100)
}
\arguments{
\item{permSTAT}{a three dimensional array with 1. = OTUs, 2. = statistics, 3. = random permutation iterations}

\item{modelratio}{1x1 scalar with the observed modelratio}

\item{trm}{1x1 scalar, at which level to truncate the OR (OR>trm := trm, OR<1/trm := 1/trm)}
}
\value{
data.frame with one row including pv (permutation p.value), SElgratio (standar error based on zero-centered normal assumptions and representability from the permutation null distribution, and some central limit theorem), permmedian (the median of the permuted null distribution), modelratio (the model ratio - as specified by input)
}
\description{
Function to extract p-values from a permutation test dataset
}
