% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2by2table.R
\name{get2by2table}
\alias{get2by2table}
\title{Calculates all 2x2 presens absense tables between two similar sized, and matched matrices}
\usage{
get2by2table(o1, o2)
}
\arguments{
\item{o1}{matrix of counts (n x p) - n is samples and p is variables}

\item{o2}{matrix of counts (n x p)}
}
\value{
a data.frame (p x 4) with the 2x2 table Ns (stats)
}
\description{
Calculates all 2x2 presens absense tables between two similar sized, and matched matrices
}
