% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistPairs.R
\name{getDistPairs}
\alias{getDistPairs}
\title{function for calculating distances between dyads}
\usage{
getDistPairs(D12, dyad)
}
\arguments{
\item{D12}{a 2n x 2n distance matric matrix}

\item{dyad}{2n x 1 vector of dyad identifiers}
}
\description{
function for calculating distances between dyads
}
