% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFisher.R
\name{getFisher}
\alias{getFisher}
\title{Calculation of OR based on Fisher exact tests and GLM}
\usage{
getFisher(x, doglm = TRUE)
}
\arguments{
\item{x}{dataset obj (1 x 4) with the 2x2 cross table of named as n00, n01, n10, n11}
}
\value{
(transfer) odds based on different methods, returned as a data.frame()
}
\description{
Calculation of OR based on Fisher exact tests and GLM
}
