% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInferenceWeightedRatio.R
\name{getInferenceWeightedRatio}
\alias{getInferenceWeightedRatio}
\title{A function to cut out inference statistics from the random permutation generated repo of sampling data generated by randpermutationTransferStats().}
\usage{
getInferenceWeightedRatio(res, trc = 100,
  include = !logical(dim(res[[1]])[1]))
}
\arguments{
\item{res}{the output from the randpermutationTransferStats}

\item{include}{logical vector indicating which of the variables to include.}
}
\value{
data.frame with model statistics and random permutation based inference statistics.
}
\description{
A function to cut out inference statistics from the random permutation generated repo of sampling data generated by randpermutationTransferStats().
}
