% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeightedRatio.R
\name{getWeightedRatio}
\alias{getWeightedRatio}
\title{Calculates the individual areas of the vulcano plot based on the OR and pv}
\usage{
getWeightedRatio(or, pv, e = 0.001)
}
\arguments{
\item{or}{Numeric vector (length k) of odd ratios}

\item{pv}{Numeric vector  (length k) of p-values}
}
\value{
dataset with np - number of positive odds (or>1), nn - number of negative odds (or<=1), ratio - the relative area between positive and negative odds overall
}
\description{
Calculates the individual areas of the vulcano plot based on the OR and pv
}
\examples{

}
