% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
geometric mean (scalar) of input
}
\description{
Calculates the geometric mean of a vector (x)
}
