% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{lag_from_names}
\alias{lag_from_names}
\title{Detect the time lags based on column names}
\usage{
lag_from_names(names, group = "taxon")
}
\description{
This is a helper function for deciding on P and Q in transfer function models
using just the names of an example set of covariates.
}
\examples{
lag_from_names(c("taxon1_lag1", "taxon1_lag2", "taxon1_lag3"))
}
