% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mbtransfer}
\alias{mbtransfer}
\title{Transfer Function Model for \code{ts_inter} objects}
\usage{
mbtransfer(
  ts_inter,
  P = 1,
  Q = 1,
  nrounds = 500,
  early_stopping_rounds = 5,
  verbose = 0,
  lambda = 0.1,
  alpha = 0.01,
  eta = 0.1,
  ...
)
}
\arguments{
\item{ts_inter}{An object of class \code{ts_inter} containing the time-varying
microbiome community, environmental interventions, and static host features
data. The columns for each element of the \code{values} matrix are expected to
be consecutive timepoints in an changing community. \code{ts_from_dfs()} is a
helper to create these objects from \code{data.frames} with the relevant
information.}

\item{P}{The number of historical community composition timepoints to
consider when making predictions.}

\item{Q}{The number of historical intervention timepoints to consider when
making predictions.}

\item{nrounds}{The maximum number of rounds for each taxon's gradient
boosting model. Smaller values will lead to faster training, but at the
risk of poorer fits. Defaults to 500.}

\item{early_stopping_rounds}{If the loss function does not improve after this
many rounds, then the model is assumed to have converged and training is
stopped. Defaults to 5.}

\item{verbose}{Should information about each gradient boosting model's
performance be printed? Allowable values are 2 (all information), 1 (some
information), and 0 (no information, default).}

\item{lambda}{The l2-regularization value in the linear gradient boosting
model. Defaults to 1e-2.}

\item{alpha}{The l1-regularization value in the linear gradient boosting
model. Defaults to 1e-2. This value generally leads to less sparse fits,
which creates useful variation in potential downstream mirror statistics
calculations.}

\item{eta}{The learning rate. Defaults to 0.05. This is slower than the
default in xgboost (0.3) but has been found to improve stability when
needing to train on taxa with a wide range of abundances.}
}
\description{
This is the main prediction function in the \code{mbtransfer} package. Given an
object of class \code{ts_inter} (see \code{ts_from_dfs()}), this will fit a collection
of linear gradient boosting-based transfer function models. The output is an
object of class \code{mbtransfer_model}. Each component boosting model is
contained in the \verb{@parameters} slot, which is a list whose j^th element is
the model for the j^th taxon (row) within each \code{ts}'s values matrix.
}
\examples{
data(sim_ts)
fit <- mbtransfer(sim_ts)
fit@parameters[[1]]
}
