% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mbtransfer_model-class}
\alias{mbtransfer_model-class}
\title{Transfer Function Model Class}
\description{
Transfer Function Model Class
}
\section{Slots}{

\describe{
\item{\code{parameters}}{The model training result. For mbtransfer models, this is a
list of all the boosting models that have been trained, one for each taxon.}

\item{\code{method}}{The model type used for prediction.}

\item{\code{hyper}}{A list. containing any hyperparameters used during model
training.}
}}

