% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mbtransfer_predict}
\alias{mbtransfer_predict}
\title{Prediction method for \code{mbtransfer} objects}
\usage{
mbtransfer_predict(object, newdata)
}
\arguments{
\item{object}{An object of class \code{mbtransfer_model}, as generated using the
\code{mbtransfer} function. This includes trained boosting models for every
taxon, stored within the \verb{@parameters} slot.}

\item{newdata}{A new \code{ts_inter} object over which to perform prediction. This
method will make predictions for every timepoint that appears in the
\verb{@interventions} slot but not the \verb{@values}.}
}
\description{
Prediction method for \code{mbtransfer} objects
}
\examples{
data(sim_ts)
fit <- mbtransfer(sim_ts)
ts_subset <- subset_values(sim_ts, 1:25)
predict(fit, ts_subset)
}
