% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBvagdevtrans}
\alias{MBvagdevtrans}
\title{Microbiome dataset on Vaginal development and Transfer to child}
\format{A phyloseq object with sample data information (sample_data())
\decribe{
\item{dyad} (the dyad number used for matching samples wihtin family)
\item{Type} (F: fecal, V: Vaginal or T: Tracheal / Airway)
\item{Time} (24: pregnany week 24, 36: pregnancy week 36, b: at birth before rupture, 1w: child age 1 week)
\item{delivery} (Normal or CSection delivery) 
\item{Lot.number} (Lot number of Extraction kit) 
\item{ExtractionTray} (The tray used for extraction)
\item{SeqRUN} (Sequncing run tray)
}}
\usage{
MBvagdevtrans
}
\description{
Microbiome dataset on Vaginal development and Transfer to child
}
\references{
Rasmussen et al (2020) ISME journal (xx.xx)
}
\keyword{datasets}
