% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_predict_single}
\alias{model_predict_single}
\title{Prediction for a Single Subject}
\usage{
model_predict_single(fit, ts_inter, lags, subject = NULL)
}
\arguments{
\item{fit}{An object of class \code{mbtransfer_model}, as generated using the
\code{mbtransfer} function. This includes trained boosting models for every
taxon, stored within the \verb{@parameters} slot.}

\item{ts_inter}{A new \code{ts_inter_single} object over which to perform
prediction. This method will make predictions for every timepoint that
appears in the \verb{@interventions} slot but not the \verb{@values}. This is assumed
to be a single subject from a larger \code{ts_inter} object.}

\item{lags}{A vector specifying \code{P} and \code{Q} in the trained mbtransfer model.}

\item{subject}{A static data frame of subject-level variables. This will be
concatenated to time-varying intervention and taxonomic covariates when
making predictions. This is analogous to the training process.}
}
\description{
This loops over all timepoints for a single subject and makes predictions by
comparing the number of timepoints in @interventions and in @values. The gap
will be filled in one step at a time using \code{mbtransfer_predict_step()}.
}
