% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{patchify_df}
\alias{patchify_df}
\title{Sliding Windows for a ts_inter object}
\usage{
patchify_df(ts_inter, p = 2, q = 3)
}
\arguments{
\item{ts_inter}{An object of class \code{ts_inter} over which to generate
sliding windows.}

\item{p}{The number of time lags to. use in the sliding window for the
microbiome features.}

\item{q}{The number of time lags to use in the sliding window for
interventions.}
}
\description{
This creates sliding windows of intervention and community composition
features for all. subjects in a \code{ts_inter} object. It returns a list giving
the "patchified" data. The first component, \code{x}, gives the microbiome and
intervention features immediately preceding the values in the second
component, \code{y}. This is constructed by running \code{patchify_single_df} over all
subjects in the dataset.
}
\examples{
data(sim_ts)
result <- patchify_df(sim_ts)
lapply(result, head)
}
