% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{patchify_single}
\alias{patchify_single}
\title{Sliding Windows for a ts_inter_single object}
\usage{
patchify_single(ts_inter, p = 2, q = 3)
}
\arguments{
\item{ts_inter}{An object of class \code{ts_inter_single} over which to generate
sliding windows.}

\item{p}{The number of time lags to. use in the sliding window for the
microbiome features.}

\item{q}{The number of time lags to use in the sliding window for
interventions.}
}
\description{
This creates sliding windows of intervention and community composition
features for a single subject. It returns a list giving the "patchified"
data, which can then be organized into matrices for prediction.
}
\examples{
library(mbtransfer)
data(sim_ts)
patchify_single(sim_ts[[1]])
}
