% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{patchify_single_df}
\alias{patchify_single_df}
\title{Convert \code{patchify_single} output into a \code{data.frame}}
\usage{
patchify_single_df(ts_inter, p, q)
}
\arguments{
\item{ts_inter}{An object of class \code{ts_inter_single} over which to generate
sliding windows.}

\item{p}{The number of time lags to. use in the sliding window for the
microbiome features.}

\item{q}{The number of time lags to use in the sliding window for
interventions.}
}
\description{
This generates a data.frame of sliding window predictors and responses by
wrapping \code{patchify_single}. It is in a format that can directly be used by
different regression methods.
}
\examples{
data(sim_ts)
patches <- patchify_single_df(sim_ts[[1]], 2, 2)
head(patches$x)
head(patches$y)
}
