% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd-mirror.R
\name{pd_effects}
\alias{pd_effects}
\title{Counterfactual Partial Dependence Effects}
\usage{
pd_effects(
  fit,
  ts,
  w0,
  w1,
  n_sample = NULL,
  patch_len = 8,
  intervention_len = NULL
)
}
\arguments{
\item{fit}{An object of class \code{mbtransfer_model}, as generated using the
\code{mbtransfer} function. This includes trained boosting models for every
taxon, stored within the \verb{@parameters} slot.}

\item{ts}{An object of class \code{ts_inter} containing the time-varying
microbiome community, environmental interventions, and static host features
data. The columns for each element of the \code{values} matrix are expected to
be consecutive timepoints in an changing community. \code{ts_from_dfs()} is a
helper to create these objects from \code{data.frames} with the relevant
information.}

\item{w0}{One of the counterfactuals with which to compute partial dependence
profiles. See \code{steps} or \code{pulses} for helpers in generating these
counterfactuals. The procedure concatenates these counterfactuals to the
end of the series and computes the difference in the forecasts.}

\item{w1}{One of the counterfactuals with which to compute partial dependence
profiles. See \code{steps} or \code{pulses} for helpers in generating these
counterfactuals. The procedure concatenates these counterfactuals to the
end of the series and computes the difference in the forecasts.}
}
\description{
Our selection algorithm, \code{select_taxa} depends on estimates for
counterfactual partial dependence effects across all taxa and time lags. This
supports estimation of these effects for a single random data split pair.
}
\examples{
data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
fit <- mbtransfer(sim_ts, 1, 1, nrounds = 10)
pd_effects(fit, sim_ts, w0, w1)
}
