% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd-mirror.R
\name{pd_splits}
\alias{pd_splits}
\title{Mirror Splits from Partial Dependence}
\usage{
pd_splits(ts, w0, w1, tr_fun, n_splits = 20, ...)
}
\description{
For ech split, we return an array of dimension n_taxa x 2 fits x n_lags. Each
entry contains the estimated counterfactual effect for that taxon and lag
combination across the two models (each fit on a different random sample of
data).
}
