% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistpair.R
\name{permDistpair}
\alias{permDistpair}
\title{Permutation function to generate null distribution of distances bewteen randomly matched microbiome, conditioning on the environment}
\usage{
permDistpair(D12, dyad, nperm = 100)
}
\arguments{
\item{D12}{a 2n x 2n distance matric matrix}

\item{dyad}{2n x 1 vector of dyad identifiers}
}
\description{
Permutation function to generate null distribution of distances bewteen randomly matched microbiome, conditioning on the environment
}
