% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict,mbtransfer_model-method}
\alias{predict,mbtransfer_model-method}
\title{Prediction method for mbtransfer models}
\usage{
\S4method{predict}{mbtransfer_model}(object, newdata)
}
\description{
Prediction method for mbtransfer models
}
