% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randpermutationTransferStats.R
\name{randpermutationTransferStats}
\alias{randpermutationTransferStats}
\title{Calculate transfer statistics including permutation testing for two phyloseq objects}
\usage{
randpermutationTransferStats(phy1, phy2, identifier, nperm = 1000,
  verbose = F)
}
\arguments{
\item{phy1}{phyloseq object for the giving environment}

\item{phy2}{phyloseq object for the recieving environment}

\item{identifier}{in sample_data for mathcing dyads(character)}
}
\description{
Calculate transfer statistics including permutation testing for two phyloseq objects
}
