% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_index.R
\name{replace_inter_}
\alias{replace_inter_}
\title{ts is a single element of a ts_inter class}
\usage{
replace_inter_(ts, new_inter, start_ix = NULL)
}
\description{
ts is a single element of a ts_inter class
}
