% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_index.R
\name{sample_ts}
\alias{sample_ts}
\title{Randomly subset a \code{ts_inter} object}
\usage{
sample_ts(ts, n, patch_len = 5, intervention_len = 0)
}
\arguments{
\item{ts}{A \code{ts_inter} object whose random patches we want to sample.}

\item{n}{The number of randomly sampled time series to return.}

\item{patch_len}{The length of the randomly sampled patches}

\item{intervention_len}{We will pad the @time slot by an extra
intervention_len set of timepoints. Defaults to 0.}
}
\description{
There are times we want to randomly sample patches from across timepoints and
subjects. This is a helper function that does this random sampling adapted to
\code{ts_inter} objects, subsetting taxa and intervention timepoints
simultaneously.
}
\examples{
data(sim_ts)
sample_ts(sim_ts, 10, 4, 1)
}
