% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_ts}
\alias{sim_ts}
\title{Simulated Time Series}
\format{
A \code{ts_inter} class of 50 subjets with 30 timepoints each.
}
\source{
This run is \code{sim_input_004.rda} from this collection,
https://drive.google.com/drive/folders/19mDoKbT2qvO9AlRdn_wQKBbqNh8Cx22F as
reshaped and interpolated by this script:
https://github.com/krisrs1128/microbiome_interventions/blob/main/scripts/forecasting_metrics.Rmd
}
\usage{
sim_ts
}
\description{
These are data generated according to a negative binomial vector
autoregressive model. The specific generative mechanism is described in
Section 3.1 of "mbtransfer: Microbiome Intervention Analysis using Transfer
Functions and Mirror Statistics" by Sankaran and Jeganathan.
}
\keyword{datasets}
