% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\name{subset_values}
\alias{subset_values}
\title{Subset values of a ts object}
\usage{
subset_values(ts, ix)
}
\description{
This is a helper to filter the number of timepoints across all \code{values} slots
in a \code{ts_inter} class.
}
\examples{
data(sim_ts)
subset_values(sim_ts, 1:2)
subset_values(sim_ts, 4:8)
}
