% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{time_lags}
\alias{time_lags}
\title{Detect the time lags over which a model was trained}
\usage{
time_lags(fit)
}
\description{
This is a helper for determining the P, Q associated with a fitted mbtransfer
model.
}
\examples{
data(sim_ts)
fit <- mbtransfer(sim_ts)
time_lags(fit)
}
