% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateZerosInf.R
\name{truncateZerosInf}
\alias{truncateZerosInf}
\title{Truncate large or small (abs(log(or)) >>> 0) odds ratios to trc or 1/trc}
\usage{
truncateZerosInf(or, trc = 100)
}
\arguments{
\item{or}{Numeric vector}
}
\value{
a vector of truncated odds ratios
}
\description{
Truncate large or small (abs(log(or)) >>> 0) odds ratios to trc or 1/trc
}
