% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ts_from_dfs}
\alias{ts_from_dfs}
\title{Build a  \code{ts}}
\usage{
ts_from_dfs(reads, interventions, metadata, subject_data = NULL)
}
\description{
Build a  \code{ts}
}
\examples{
library(readr)
library(tibble)
subject <- read_figshare_csv("https://ndownloader.figshare.com/files/40275934")
interventions <- read_figshare_csv("https://ndownloader.figshare.com/files/40279171") |>
  column_to_rownames("sample")
reads <- read_figshare_csv("https://ndownloader.figshare.com/files/40279108") |>
  column_to_rownames("sample")
samples <- read_figshare_csv("https://ndownloader.figshare.com/files/40275943")
ts <- as.matrix(reads) |>
  ts_from_dfs(interventions, samples, subject)
ts
}
