% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\docType{class}
\name{ts_inter-class}
\alias{ts_inter-class}
\title{An S4 class representing a collection of time series under environmental shifts}
\description{
While \code{ts_inter_single} represents shifts for a single subject, \code{ts_inter}
includes shifts across a collection of subjects. It includes an additional
slot, \code{subject_data} to store non-time-varying subject metadata.
}
\section{Slots}{

\describe{
\item{\code{series}}{A list of \code{ts_inter_single} objects, each representing
interventions and compositional responses for a subject.}

\item{\code{subject_data}}{An optional data.frame storing static host-level metadata
associated with each series.}
}}

\examples{
data(sim_ts)
sim_ts
}
