% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\docType{class}
\name{ts_inter_single-class}
\alias{ts_inter_single-class}
\title{An S4 class to represent a single subject's series}
\description{
Taxonomic abundances, binary/continuous interventions, and underlying
timepoints for a single subject are stored in this unified data structure. A
\code{ts_inter} object is list of these \code{ts_inter_single} classes (together wtih
static subject data).
}
\section{Slots}{

\describe{
\item{\code{values}}{A matrix whose rows are taxa and whose columns are samples.}

\item{\code{time}}{A vector of the timepoints associated with the samples in \code{values}.}

\item{\code{interventions}}{A matrix whose rows are perturbations, columns are
samples, and values are either binary interventions or continuous input
series, representing the value of the exogenous influence.}
}}

\examples{
data(sim_ts)
sim_ts[[1]]
}
