% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_generation.R
\name{counterfactual_ts}
\alias{counterfactual_ts}
\title{Generate Counterfactual versions of a ts_inter object}
\usage{
counterfactual_ts(ts, w0, w1, start_ix = NULL)
}
\arguments{
\item{ts}{The starting ts_inter object from which to build the '
counterfactuals. By default, this is the last timepoint in the current series.}

\item{w0}{The first version of the intervention series to consider.}

\item{w1}{The second version of the intervention series to consider.}

\item{start_ix}{The truncation position for the original ts. Defaults to no
truncation, which appends new interventions to the end of the existing
series.}
}
\value{
Two ts_inter objects with interventions corresponding to w0 and w1.
}
\description{
mbtransfer makes predictions starting from ts_inter objects, so in order to
simulate counterfactuals we need to provide alternative versions of those
objects as input. This function truncates an existing ts object at start_ix
and creates new intervention series according to the values of w0 and w1
(starting from the truncation point).
}
\examples{
data(sim_ts)
ws <- steps(c("P1" = TRUE), 2, 3, 4)
ts_star <- counterfactual_ts(sim_ts, ws[[1]], ws[[2]], start_ix = 10)
}
