% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd-mirror.R
\name{pd_splits}
\alias{pd_splits}
\title{Mirror Splits from Partial Dependence}
\usage{
pd_splits(ts, w0, w1, tr_fun, n_splits = 20, ...)
}
\arguments{
\item{ts}{An object of class \code{ts_inter} containing the time-varying
microbiome community, environmental interventions, and static host features
data. The columns for each element of the \code{values} matrix are expected to
be consecutive timepoints in an changing community. \code{ts_from_dfs()} is a
helper to create these objects from \code{data.frames} with the relevant
information.}

\item{w0}{One of the counterfactuals with which to compute partial dependence
profiles. See \code{steps} or \code{pulses} for helpers in generating these
counterfactuals. The procedure concatenates these counterfactuals to the
end of the series and computes the difference in the forecasts.}

\item{w1}{One of the counterfactuals with which to compute partial dependence
profiles. See \code{steps} or \code{pulses} for helpers in generating these
counterfactuals. The procedure concatenates these counterfactuals to the
end of the series and computes the difference in the forecasts.}

\item{tr_fun}{A function that can be used to train over random splits. In the
examples in this package, we use an anonymous function that fills our
chosen hyperparameters in \code{mbtransfer}. For example \verb{\\(x) mbtransfer(x, P =  1, Q = 1)}
will fit a lag-1 transfer function model on all the random splits.}
}
\description{
For ech split, we return an array of dimension n_taxa x 2 fits x n_lags. Each
entry contains the estimated counterfactual effect for that taxon and lag
combination across the two models (each fit on a different random sample of
data).
}
\examples{
data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
pd_result <- pd_splits(sim_ts, w0, w1, \(x) mbtransfer(x, 1, 1, nrounds = 10), n_splits = 1)
}
