% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict,mbtransfer_model-method}
\alias{predict,mbtransfer_model-method}
\title{Prediction method for mbtransfer models
update method}
\usage{
\S4method{predict}{mbtransfer_model}(object, newdata)
}
\description{
Prediction method for mbtransfer models
update method
}
\examples{
data(sim_ts)
fit <- mbtransfer(sim_ts, 1, 1)
sim_sub <- subset_values(sim_ts, 1:26) # remove last timepoint
y_hat <- predict(fit, sim_sub)
plot(values(sim_ts[[1]])[, 27], values(y_hat[[1]])[, 27], xlab = "Truth", ylab = "Predicted")
}
